+++ src/lib/components/forms/form/form-components/double-input/double-input.interface.ts
@@ -1,0 +1,4 @@
+export interface DoubleInputInterface {
+    isPriceInput: boolean;
+    decimalCount: number;
+}